Move port metadata file to process orchestrator create
