+++ src/orchestrator-process/Cargo.toml
@@ -20,3 +20,6 @@ scopeguard = "1.1.0"
 sysinfo = "0.23.10"
 tokio = "1.17.0"
 tracing = "0.1.34"
+
+[dev-dependencies]
+tempfile = "3.2.0"
