+++ src/orchestrator-process/src/lib.rs
@@ -7,7 +7,7 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
-mod port_metadata_file;
+pub mod port_metadata_file;
 
 use std::collections::HashMap;
 use std::fmt::Debug;
